+++ source/runner/_tests/_executor.test.ts
@@ -3,13 +3,13 @@ import { FakeCI } from "../../ci_source/providers/Fake"
 import { FakePlatform } from "../../platforms/FakePlatform"
 import {
   emptyResults,
-  warnResults,
-  inlineWarnResults,
   failsResults,
   inlineFailResults,
   inlineMessageResults,
   inlineMultipleWarnResults,
   inlineMultipleWarnResults2,
+  inlineWarnResults,
+  warnResults,
 } from "./fixtures/ExampleDangerResults"
 import inlineRunner from "../runners/inline"
 import { jsonDSLGenerator } from "../dslGenerator"
@@ -17,7 +17,7 @@ import { jsonToDSL } from "../jsonToDSL"
 import { DangerDSLType } from "../../dsl/DangerDSL"
 import { singleViolationSingleFileResults } from "../../dsl/_tests/fixtures/ExampleDangerResults"
 import { inlineTemplate } from "../templates/githubIssueTemplate"
-import { resultsIntoInlineResults, DangerResults, inlineResultsIntoResults } from "../../dsl/DangerResults"
+import { DangerResults, inlineResultsIntoResults, resultsIntoInlineResults } from "../../dsl/DangerResults"
 
 const defaultConfig: ExecutorOptions = {
   stdoutOnly: false,
@@ -27,6 +27,7 @@ const defaultConfig: ExecutorOptions = {
   passURLForDSL: false,
   failOnErrors: false,
   noPublishCheck: false,
+  ignoreOutOfDiffComments: false,
 }
 
 class FakeProcces {
@@ -120,6 +121,7 @@ describe("setup", () => {
       dangerID: "123",
       passURLForDSL: false,
       failOnErrors: true,
+      ignoreOutOfDiffComments: false,
     }
     const exec = new Executor(new FakeCI({}), platform, inlineRunner, strictConfig, new FakeProcces())
     const dsl = await defaultDsl(platform)
@@ -172,10 +174,7 @@ describe("setup", () => {
     const platform = new FakePlatform()
     const exec = new Executor(new FakeCI({}), platform, inlineRunner, defaultConfig, new FakeProcces())
     const dsl = await defaultDsl(platform)
-    const apiFailureMock = jest.fn().mockReturnValue(
-      new Promise<any>((_, reject) => reject())
-    )
-    platform.createInlineComment = apiFailureMock
+    platform.createInlineComment = jest.fn().mockReturnValue(new Promise<any>((_, reject) => reject()))
 
     let results = await exec.sendInlineComments(singleViolationSingleFileResults, dsl.git, [])
     expect(results).toEqual(singleViolationSingleFileResults)
@@ -190,6 +189,18 @@ describe("setup", () => {
 
     await exec.handleResults(inlineWarnResults, dsl.git)
     expect(platform.createInlineComment).toBeCalled()
+  })
+
+  it("Invalid inline comment is ignored if ignoreOutOfDiffComments is true", async () => {
+    const platform = new FakePlatform()
+    const config = Object.assign({}, defaultConfig, { ignoreOutOfDiffComments: true })
+    const exec = new Executor(new FakeCI({}), platform, inlineRunner, config, new FakeProcces())
+    const dsl = await defaultDsl(platform)
+    platform.createInlineComment = jest.fn().mockReturnValue(new Promise<any>((_, reject) => reject()))
+    platform.updateOrCreateComment = jest.fn()
+
+    await exec.handleResults(inlineWarnResults, dsl.git)
+    expect(platform.updateOrCreateComment).not.toBeCalled()
   })
 
   it("Updates an inline comment as the new one was different than the old", async () => {
@@ -270,10 +281,7 @@ describe("setup", () => {
     const dsl = await defaultDsl(platform)
     const previousResults = {
       fails: [],
-      warnings: [
-        { message: "1", file: "1.swift", line: 1 },
-        { message: "2", file: "2.swift", line: 2 },
-      ],
+      warnings: [{ message: "1", file: "1.swift", line: 1 }, { message: "2", file: "2.swift", line: 2 }],
       messages: [],
       markdowns: [],
     }
@@ -297,19 +305,13 @@ describe("setup", () => {
     const dsl = await defaultDsl(platform)
     const previousResults = {
       fails: [],
-      warnings: [
-        { message: "1", file: "1.swift", line: 1 },
-        { message: "2", file: "2.swift", line: 2 },
-      ],
+      warnings: [{ message: "1", file: "1.swift", line: 1 }, { message: "2", file: "2.swift", line: 2 }],
       messages: [],
       markdowns: [],
     }
     const newResults = {
       fails: [],
-      warnings: [
-        { message: "1", file: "1.swift", line: 2 },
-        { message: "2", file: "2.swift", line: 3 },
-      ],
+      warnings: [{ message: "1", file: "1.swift", line: 2 }, { message: "2", file: "2.swift", line: 3 }],
       messages: [],
       markdowns: [],
     }