+++ CHANGELOG.md
@@ -15,6 +15,9 @@
 
 <!-- Your comment below this -->
 
+- Added a CLI option `--ignoreOutOfDiffComments` so that you can ignore inline-comments for lines that were not changed
+  in the checked PR. The comments would be ignored completely - they won't even show in the results comment. [@pinkasey]
+
 <!-- Your comment above this -->
 
 # 10.2.2